pyproject.toml
[project]
name = "enterprise-incident-manager"
version = "0.1.0"
description = "A production-ready FastAPI Incident Management System"
authors = [
    {name = "Rishi Kant", email = "x7rishi@gmail.com"}
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.2.0",
    "alembic>=1.13.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0", # For testing
    "elasticsearch>=8.0.0,<9.0.0",
    "aiohttp", 
    "bcrypt==4.3.0"
]
requires-python = ">=3.11"

[project.optional-dependencies]
dev = [
    "black>=0.3.0",     # Linter and Formatter (Senior choice)
    "mypy>=1.9.0",     # Static Type Checker
    "pytest>=8.0.0",   # Testing framework
    "pytest-asyncio>=0.23.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]  # Only include the 'app' directory and its sub-packages
exclude = ["tests*", "migrations*"] # Ignore tests and migrations during installation